+++ graphbrain/constants.py
@@ -1,4 +1,6 @@
 # Pre-defined system atoms
+sequence_head_pred = 'head/p/.'
+sequence_next_pred = 'next/p/.'
 noun_connector_pred = '+/b/.'
 lemma_pred = 'lemma/p/.'
 coref_pred = 'coref/p/.'